+++ packages/databyss-ui/editor/EditorInline.js
@@ -1,131 +1,21 @@
 import React, { useState, useEffect } from 'react'
 import { color } from 'styled-system'
 import styled from '@emotion/styled'
-import {
-  View,
-  Button,
-  Text,
-  Grid,
-  Icon,
-  ModalWindow,
-  TextControl,
-  List,
-} from '@databyss-org/ui/primitives'
+import { View, Button, Icon } from '@databyss-org/ui/primitives'
 import buttons from '@databyss-org/ui/theming/buttons'
 import Close from '@databyss-org/ui/assets/angle-right-solid.svg'
 import { updateSource } from './state/page/actions'
 import { useEditorContext } from './EditorProvider'
-import {
-  useSourceContext,
-  withSource,
-} from '@databyss-org/services/sources/SourceProvider'
-import ValueListProvider, {
-  ValueListItem,
-} from '@databyss-org/ui/components/ValueList/ValueListProvider'
+import SourceModal from '@databyss-org/ui/modules/SourcesValueList/SourceModal'
 
 const Styled = styled('span')(color)
 
 const { buttonVariants } = buttons
 
-const modal = {
-  title: 'Edit Source',
-  dismissChild: 'Save',
-  secondaryChild: 'Cancel',
-}
-
-const ControlList = ({ children, ...others }) => (
-  <List horizontalItemPadding="small" {...others}>
-    {children}
-  </List>
-)
-
-const SourceModal = withSource(
-  ({ source, visible, setVisible, onUpdateSource, ...modal }) => {
-    const [values, setValues] = useState(source)
-    const [, setSource] = useSourceContext()
-
-    const onChange = _value => {
-      // update internal state
-      setValues(_value)
-    }
-
-    const onSave = () => {
-      setVisible(false)
-      setSource(values)
-      onUpdateSource(values)
-      // set internal dispatch here
-    }
-
-    return (
-      <ModalWindow visible={visible} onDismiss={() => onSave()} {...modal}>
-        <ValueListProvider onChange={onChange} values={values}>
-          <Grid>
-            <View
-              paddingVariant="none"
-              widthVariant="content"
-              backgroundColor="background.0"
-              width="100%"
-            >
-              <ControlList verticalItemPadding="tiny">
-                <ValueListItem path="text">
-                  <TextControl
-                    labelProps={{
-                      width: '25%',
-                    }}
-                    label="Name"
-                    id="name"
-                    gridFlexWrap="nowrap"
-                    paddingVariant="tiny"
-                    rich
-                  />
-                </ValueListItem>
-                <ValueListItem path="authors[0].firstName">
-                  <TextControl
-                    labelProps={{
-                      width: '25%',
-                    }}
-                    label="Author (First Name)"
-                    id="firstName"
-                    gridFlexWrap="nowrap"
-                    paddingVariant="tiny"
-                  />
-                </ValueListItem>
-                <ValueListItem path="authors[0].lastName">
-                  <TextControl
-                    labelProps={{
-                      width: '25%',
-                    }}
-                    label="Author (Last Name)"
-                    id="lastName"
-                    gridFlexWrap="nowrap"
-                    paddingVariant="tiny"
-                  />
-                </ValueListItem>
-                <ValueListItem path="citations[0]">
-                  <TextControl
-                    labelProps={{
-                      width: '25%',
-                    }}
-                    label="Citation"
-                    id="citation"
-                    rich
-                    gridFlexWrap="nowrap"
-                    multiline
-                    paddingVariant="tiny"
-                  />
-                </ValueListItem>
-              </ControlList>
-            </View>
-          </Grid>
-        </ValueListProvider>
-      </ModalWindow>
-    )
-  }
-)
-
 const EditorInline = React.forwardRef(
   ({ backgroundColor, node, children, ...others }, ref) => {
     const [visible, setVisible] = useState(false)
+
     const [refId, setRefId] = useState(null)
 
     const [editorState, dispatchEditor] = useEditorContext()
@@ -148,7 +38,6 @@ const EditorInline = React.forwardRef(
       const _idList = Object.keys(blocks).filter(
         (block, i) => blocks[block].refId === source._id
       )
-
       dispatchEditor(
         updateSource(source, _idList, { value: editableState.value })
       )
@@ -159,29 +48,23 @@ const EditorInline = React.forwardRef(
         {children}
         <View display="inline-block">
           <Button
-            backgroundColor={backgroundColor}
+            variant="editSource"
             onClick={() => setVisible(true)}
             data-test-atomic-edit="open"
           >
-            <Icon
-              sizeVariant="tiny"
-              color={buttonVariants.editorMarginMenu.color}
-            >
+            <Icon sizeVariant="tiny" color="background.5">
               <Close />
             </Icon>
           </Button>
         </View>
-        <View>
-          {refId && (
-            <SourceModal
-              sourceId={refId}
-              visible={visible}
-              setVisible={setVisible}
-              onUpdateSource={onUpdateSource}
-              {...modal}
-            />
-          )}
-        </View>
+        {refId && (
+          <SourceModal
+            sourceId={refId}
+            visible={visible}
+            setVisible={setVisible}
+            onUpdateSource={onUpdateSource}
+          />
+        )}
       </Styled>
     )
   }
